Add files via upload
Context:
DeltaSquare is an NGO that works with the Government on matters of social policy to bring about a change in the lives of underprivileged sections of society. They are tasked with coming up with a policy framework by looking at the data government got from WHO. You as a data scientist at DeltaSquare are tasked with solving this problem and sharing a proposal for the government.
 

Problem:
The dataset aims to answer the following key questions:

What are the different factors that influence the income of an individual?
Is there a good predictive model for income that exists? 
What does the performance assessment look like for such a model?